+++ lib/domain/entities/recipe/recipe_entity.dart
@@ -60,7 +60,7 @@ class RecipeApiEntity {
   final dynamic preparationMinutes;
   final dynamic cookingMinutes;
   final int? aggregateLikes;
-  final int? healthScore;
+  final double? healthScore;
   final String? creditsText;
   final String? license;
   final String? sourceName;
